+++ ICD.Connect.Cameras.Windows/Properties/AssemblyInfo.cs
@@ -1,0 +1,7 @@
+using System.Reflection;
+
+[assembly: AssemblyTitle("ICD.Connect.Cameras.Windows")]
+[assembly: AssemblyCompany("ICD Systems")]
+[assembly: AssemblyProduct("ICD.Connect.Cameras.Windows")]
+[assembly: AssemblyCopyright("Copyright © ICD Systems 2019")]
+[assembly: AssemblyVersion("4.2.0.0")]