+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+ - Initial commit of ICD.Connect.Cameras.Windows project
+ - Added WindowsUsbCameraDevice for representing static USB cameras
+
 ## [4.2.0] - 2019-11-18
 ### Added
  - Added web proxy settings to Panasonic camera driver